refactor: extract MouseEventHandler from RichLogViewer
Move all mouse event handling logic (click, drag, double-click) into a
dedicated MouseEventHandler class.
